+++ PoolLightPie/PoolLightPie/Services/Interfaces/IServiceLectureTemperature.cs
@@ -2,6 +2,6 @@
 {
     public interface IServiceLectureTemperature
     {
-        double Lire();
+        float Lire();
     }
 }
